LCD/LED tweaks, 2 analog inputs for mode selection & boot config options

+++ src/LEDControl.h
@@ -5,12 +5,12 @@
 
 class LEDControl{
     
-    int strength;
+    float strength;
     
     
     public:
         //LEDControl();
-        void update(int);
+        void update(float);
     
     };
